django/requirements.txt
asgiref==3.2.7
astroid==2.4.2
cffi==1.14.0
confusable-homoglyphs==3.2.0
cryptography==2.9.2
Django==3.0.7
django-fluent-contents==2.0.7
django-fluent-utils==2.0.1
django-geojson==3.0.0
django-geojson-serializer==0.2
django-mapbox-location-field==1.5.0
django-parler==2.0.1
django-polymorphic==2.1.2
django-registration==3.1
django-rest-knox==4.1.0
django-tag-parser==3.2
django-template-analyzer==1.6.2
djangorestframework==3.11.0
djangorestframework-gis==0.15
future==0.18.2
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.5
pycparser==2.20
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
webencodings==0.5.1
wrapt==1.12.1
